+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 - added `services_with_errors` count to the `/api/overall_status` endpoint
+- added new endpoint `/api/check_now/<id>` to reset a given hosts next check time to now
 
 ### Changed